% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForExposure.R
\name{checkForExposure}
\alias{checkForExposure}
\title{Check if exposure exists inside an ExposomeSet}
\usage{
checkForExposure(set, exposure, datasources = NULL)
}
\arguments{
\item{set}{\code{character} Name of the ExposomeSet on the study server}

\item{exposure}{\code{character} Name of the exposure to look for on the ExposomeSet}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
No return object, only stops the execution if the exposure is not on the ExposomeSet
}
\description{
Internal function
}
