% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_corPlot.R
\name{circos_corPlotEXP}
\alias{circos_corPlotEXP}
\title{Draws correlation circos plot}
\usage{
circos_corPlotEXP(
  corr,
  desc,
  correlation.limits = list(list(d = "+", t = 0.5, c = "#191970"), list(d = "+", t =
    0.3, c = "#4169E1"), list(d = "-", t = -0.3, c = "#DC143C"), list(d = "-", t = -0.5,
    c = "#8B0000")),
  show.legend = TRUE,
  cex.exposures = 0.5,
  cex.family = 0.55,
  colors
)
}
\arguments{
\item{corr}{Correlation}

\item{desc}{Fams}

\item{correlation.limits}{Limits}

\item{show.legend}{Bool}

\item{cex.exposures}{Alpha}

\item{cex.family}{Alpha}

\item{colors}{Cols}
}
\description{
This is a internal function, the code itself is taken from rexposome::.correlation_circos_plot
}
