% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPlot.R
\name{corPlot}
\alias{corPlot}
\title{Draws correlation plot}
\usage{
corPlot(corr_list, type = "matrix")
}
\arguments{
\item{corr_list}{\code{list} Output of \code{\link{ds.exposome_correlation}}}

\item{type}{\code{character} Type of the plot, \code{"matrix"} for a matrix plot or \code{"circos"} for a circos
plot}
}
\value{
Renders the selected plot on the default graphical device
}
\description{
Takes the output of the \code{\link{ds.exposome_correlation}} function and draws a plot (circos or matrix)
}
\examples{
\dontrun{Refer to the package Vignette for examples.}
}
