% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.anderson.darling.test.R
\name{ds.anderson.darling.test}
\alias{ds.anderson.darling.test}
\title{Anderson-Darling Normality Test}
\usage{
ds.anderson.darling.test(x, datasources = NULL)
}
\arguments{
\item{x}{\code{character} Name of the numeric vector on the server side}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
Results of the Anderson-Darling Normality Test
}
\description{
Perform the Anderson-Darling Normality Test on a numeric vector (or column of a data frame)
}
