% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.exposome_summary.R
\name{ds.exposome_summary}
\alias{ds.exposome_summary}
\title{Generate the summary of a variable of an exposome dataset}
\usage{
ds.exposome_summary(Set, variable, datasources = NULL)
}
\arguments{
\item{Set}{\code{character} Name of the Exposome Set object on the server side}

\item{variable}{\code{character} Name of the variable to obtain it's summary (numeric or factor variable)}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
The function returns a \code{list}, the contents of it depend on the variable nature, for numeric variables: \cr
- \code{$class} Class of the variable \cr
- \code{$length} Number of observations of the variable \cr
- \code{$quantiles & mean} Quantiles and mean of the variable \cr

Categorical variable: \cr
- \code{$class} Class of the variable \cr
- \code{$length} Number of observations of the variable \cr
- \code{$categories} Levels of the categorical variable \cr
- \code{$counts of each level} Counts for each of the levels
}
\description{
Generates the summary of a variable (phenotype or exposure) of a server side 'ExposomeSet' object
}
\examples{
\dontrun{Refer to the package Vignette for examples.}
}
