% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.exposome_variables.R
\name{ds.exposome_variables}
\alias{ds.exposome_variables}
\title{Get names from the exposures or phenotypes of an Exposome Set of the server side}
\usage{
ds.exposome_variables(Set, target = "exposures", datasources = NULL)
}
\arguments{
\item{Set}{\code{character} Name of the Exposome Set object on the server side}

\item{target}{\code{character} To specify target of output table, \code{exposures} for the exposure names,
\code{phenotypes} for the phenotypes. Default \code{exposures}}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
A \code{list} with the variable names selected
}
\description{
Extract all the different variables of the phenotypes or exposures of an exposome dataset on the
server side
}
\examples{
\dontrun{Refer to the package Vignette for examples.}
}
