% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.exposomeSubset.R
\name{ds.exposomeSubset}
\alias{ds.exposomeSubset}
\title{Subset Exposome Set by familie(s)}
\usage{
ds.exposomeSubset(set, fam, name = NULL, datasources = NULL)
}
\arguments{
\item{set}{\code{character} Name of the Exposome Set object on the server side}

\item{fam}{\code{character vector} Families to subset the exposome set}

\item{name}{\code{character} (default \code{NULL}) Name of the new Exposome Set, if null the name 
will be \code{"set_subsetted"} where \code{"set"} is the inputted argument}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
This function does not have an output. It creates an Exposome Set object on the study server.
}
\description{
Subset an Exposome Set on the server side by familie(s), if no \code{name} argument is provided, 
the new Exposome Set object will be named \code{"set_subsettted"} where \code{"set"} is the inputted argument
}
