% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.familyNames.R
\name{ds.familyNames}
\alias{ds.familyNames}
\title{Get family names from Exposome Set}
\usage{
ds.familyNames(object, by.exposure = FALSE, datasources = NULL)
}
\arguments{
\item{object}{\code{character} Name of the Exposome Set object on the server side}

\item{by.exposure}{\code{bool} (default \code{FALSE}) If \code{TRUE} a vector labeled with each exposure 
name will be returned with the family of each exposures. If \code{FALSE} a vector with the 
(unique) name of the families of exposures will be returned.}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
A \code{character vector} with the family names of the Exposome Set
}
\description{
Get family names from Exposome Set on the server side
}
\examples{
\dontrun{Refer to the package Vignette for examples.}
}
