% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.get_exposure_from_geo.R
\name{ds.get_exposure_from_geo}
\alias{ds.get_exposure_from_geo}
\title{Associate `NetCDF` data and individual location}
\usage{
ds.get_exposure_from_geo(
  lat,
  lon,
  exposure,
  exposure_name,
  clinical,
  clinical_lat_var,
  clinical_lon_var,
  clinical_id_var,
  exposures = NULL,
  exposures_id_var = NULL,
  new.obj = "exposures",
  datasources = NULL
)
}
\arguments{
\item{lat}{`character` (Object server name of: ) 
Latitude extracted from a `NetCDF` resource using `ncvar_getDS`}

\item{lon}{`character` (Object server name of: ) 
Longitutde extracted from a `NetCDF` resource using `ncvar_getDS`}

\item{exposure}{`character` (Object server name of: ) 
Exposure extracted from a `NetCDF` resource using `ncvar_getDS`}

\item{exposure_name}{`character` Name to assign to the `exposure`}

\item{clinical}{`character` (Object server name of: ) 
 Clinical data data of the individuals, which contain among
other variables, the longitude and latitude of the individuals to be associated with the 
`NetCDF` data.}

\item{clinical_lat_var}{`character` Name of the latitude variable on the `clinical` data}

\item{clinical_lon_var}{`character` Name of the longitude variable on the `clinical` data}

\item{clinical_id_var}{`character` Name of the IDs variable on the `clinical` data}

\item{exposures}{(default `NULL`) (Object server name of: ) 
 If provided, existing `data.frame` with exposure data.
The new exposure calculated from the `NetCDF` data will be added to this table by ID}

\item{exposures_id_var}{(default `NULL`) Required if `exposures` is provided. Name of the
IDs variable on the `exposures` data.}
}
\value{
Creates object on the server
}
\description{
Merge the exposure data contained in a `NetCDF` object
with the location data of the individuals present on a `clinical` `data.frame`
}
