% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.imputation.R
\name{ds.imputation}
\alias{ds.imputation}
\title{Perform a imputation of an Exposome Set}
\usage{
ds.imputation(object, name = NULL, select = NULL, datasources = NULL)
}
\arguments{
\item{object}{\code{character} Name of the Exposome Set on the server side}

\item{name}{\code{character} (default \code{NULL}) Name to be assigned to the exposed set on the server side. If missing,
the input Exposome Set will be over written.}

\item{select}{\code{character} (default \code{NULL}) Exposures to be imputed. If missing, all exposes will be imputed.}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
This function does not have an output. It creates an Exposome Set object on the study server.
}
\description{
Perform a imputation of an Exposome Set on the server side using rexposome base function
}
\examples{
\dontrun{Refer to the package Vignette for examples.}
}
