% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.invExWAS.R
\name{ds.invExWAS}
\alias{ds.invExWAS}
\title{Title}
\usage{
ds.invExWAS(object, formula, tef = TRUE, datasources = NULL)
}
\arguments{
\item{object}{}

\item{formula}{}

\item{tef}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
