% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.kolmogorov.smirnov.test.R
\name{ds.kolmogorov.smirnov.test}
\alias{ds.kolmogorov.smirnov.test}
\title{Kolmogorov-Smirnov Normality Test}
\usage{
ds.kolmogorov.smirnov.test(x, datasources = NULL)
}
\arguments{
\item{x}{\code{character} Name of the numeric vector on the server side}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
Results of the Kolmogorov-Smirnov Normality Test
}
\description{
Perform the Kolmogorov-Smirnov Normality Test on a numeric vector (or column of a data frame)
}
