% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.loadExposome.R
\name{ds.loadExposome}
\alias{ds.loadExposome}
\title{Creation of an ExposomeSet from \code{data.frames}}
\usage{
ds.loadExposome(
  exposures,
  description,
  phenotypes,
  description.famCol = "family",
  exposures.asFactor = 5,
  warnings = FALSE,
  object_name = "exposome_set",
  datasources = NULL
)
}
\arguments{
\item{exposures}{\code{character} Name of the exposures variable on the study server}

\item{description}{\code{character} Name of the description variable on the study server}

\item{phenotypes}{\code{character} Name of the phenotypes variable on the study server}

\item{description.famCol}{(default \code{"family"}) Index where the family's
name (per exposures) if found in file "description". It can be both numeric
or character.}

\item{exposures.asFactor}{\code{numeric} (default \code{5}) The exposures with more
than this number of unique items will be considered as "continuous" while
the exposures with less or equal number of items will be considered as
"factor".}

\item{warnings}{\code{bool} (default \code{FALSE}) If \code{TRUE} shows useful
information/warnings from the process of loading the exposome.}

\item{object_name}{\code{character} (default \code{"exposome_set"}) Name assigned to the Exposome Set on the study server}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
This function does not have an output. It creates an Exposome Set object on the study server.
}
\description{
Takes the three tables of an exposome dataset and coerces them into a Exposome Set object
on the study server
}
\examples{
\dontrun{Refer to the package Vignette for examples.}

}
