% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.maturityoffset.R
\name{ds.maturityoffset}
\alias{ds.maturityoffset}
\title{Maturity offset estimation (pubertal development)}
\usage{
ds.maturityoffset(data, sex, age, height, new.obj = NULL, datasources = NULL)
}
\arguments{
\item{data}{Dataset name (it must be a data frame)}

\item{sex}{Name of the sex variable in the input dataset (0 for males and 1 for females)}

\item{age}{Name of the Age variable in the input dataset (in days)}

\item{height}{Name of the Height variable in the input dataset (in centimeters)}

\item{new.obj}{Name of the output object containing maturity offset estimations (default: "var_maturityoffset")}

\item{datasources}{details for DS connections}
}
\value{
Estimated maturity offset according to Moore's equations.
}
\description{
Calculate maturity offset according to the equations of Moore et al. 2015 (10.1249/MSS.0000000000000588)
}
\details{
More details can be found at Moore et al. 2015 (10.1249/MSS.0000000000000588)
}
\author{
Augusto Anguita-Ruiz & Xavier Escribá
}
