% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.ncvar_get.R
\name{ds.ncvar_get}
\alias{ds.ncvar_get}
\title{Title}
\usage{
ds.ncvar_get(nc, varid, new.obj = "ncvar", datasources = NULL)
}
\arguments{
\item{nc}{}

\item{varid}{}

\item{new.obj}{}

\item{datasources}{}
}
\description{
Title
}
