% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.NetCDF_fillvalue_matrix.R
\name{ds.NetCDF_fillvalue_matrix}
\alias{ds.NetCDF_fillvalue_matrix}
\title{Set NAs to matrix fillvalues}
\usage{
ds.NetCDF_fillvalue_matrix(mat, fillvalue, new.obj = "mat", datasources = NULL)
}
\arguments{
\item{mat}{(Object server name of: ) `matrix` Extracted from a `NetCDF` resource using `nc_dataDS`}

\item{fillvalue}{(Object server name of: ) `list` Extracted from the `_FillValue` slot of a `NetCDF`
resource using `ncatt_getDS`}
}
\value{
Creates object on the server
}
\description{
Use a fillvalue extracted using `ncatt_get` and place NAs
}
