% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.netcdf_vars.R
\name{ds.netcdf_vars}
\alias{ds.netcdf_vars}
\title{Get variable names of NetCDF object}
\usage{
ds.netcdf_vars(nc, datasources = NULL)
}
\arguments{
\item{nc}{(Object server name of: )`character`
An object of class ncdf4 (loaded from a NetCDF resource), 
indicating what file to get the variables from.}
}
\value{
`character vector` With the variable names
}
\description{
Get variable names of NetCDF object
}
