% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.normalityTest.R
\name{ds.normalityTest}
\alias{ds.normalityTest}
\title{Normality test of exposures of ExposomeSet}
\usage{
ds.normalityTest(
  object,
  th = 0.05,
  min.val = 5,
  na.rm = TRUE,
  warnings = TRUE,
  datasources = NULL
)
}
\arguments{
\item{object}{\code{character} Name of the Exposome Set on the server side}

\item{th}{\code{numeric} (default \code{0.05}) Threshold to considere an exposure to follow a normal distribution.}

\item{min.val}{\code{numeric} (default \code{5}) Minimum number of values not missings to test the exposures.}

\item{na.rm}{\code{bool} (default \code{TRUE}) Removes the NA values to test the normality on the exposure.}

\item{warnings}{\code{bool} (default \code{TRUE}) Show warnings if required.}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
list of \code{data.frame} (one for each study server) with \cr
- exposure: \code{character} Name of the exposures \cr
- normality: \code{bool} If exposure is normal (\code{TRUE}) or not (\code{FALSE}) \cr
- p.value: \code{numeric} P-value of the Shapiro-Wilk Normality Test
}
\description{
Perform a normality test on all the exposures of an ExposomeSet object of the server side
}
\examples{
\dontrun{Refer to the package Vignette for examples.}
}
