% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.phenotypeNames.R
\name{ds.phenotypeNames}
\alias{ds.phenotypeNames}
\title{Exposure Phenotype names}
\usage{
ds.phenotypeNames(x, datasources = NULL)
}
\arguments{
\item{x}{\code{character} ExposomeSet object}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}
}
\value{
\code{character vector} of the phenotypes names
}
\description{
Extract the phenotypes names from a Expression Set object
}
