% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.shapiro.test.R
\name{ds.shapiro.test}
\alias{ds.shapiro.test}
\title{Shapiro-Wilk Normality Test}
\usage{
ds.shapiro.test(x, datasources = NULL)
}
\arguments{
\item{x}{\code{character} Name of the numeric vector on the server side}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
Results of the Shapiro-Wilk Normality Test
}
\description{
Perform the Shapiro-Wilk Normality Test on a numeric vector (or column of a data frame)
}
