% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.standardize.R
\name{ds.standardize}
\alias{ds.standardize}
\title{Title}
\usage{
ds.standardize(set, method = "normal", name = NULL, datasources = NULL)
}
\arguments{
\item{set}{}

\item{method}{}

\item{name}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
