% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.subset_type.R
\name{ds.subset_type}
\alias{ds.subset_type}
\title{Subset data frame by class}
\usage{
ds.subset_type(data, type = "numeric", newobj = NULL, datasources = NULL)
}
\arguments{
\item{data}{\code{character} Name of the data frame on the study server}

\item{type}{\code{character} Data type to which subset the data frame}

\item{newobj}{\code{character} Name of the output object assigned to the server side, if \code{NULL} the input table 
will be over-written}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}
}
\value{
This function does not have an output. It creates an Exposome Set object on the study server.
}
\description{
Create a subset of a data frame with the columns that have a shared class
}
