% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tableMissings.R
\name{ds.tableMissings}
\alias{ds.tableMissings}
\title{Get table of missings per variable}
\usage{
ds.tableMissings(exp, set = "exposures", output = "n", datasources = NULL)
}
\arguments{
\item{exp}{\code{character} Name of the Exposome Set on the server side}

\item{set}{\code{character} (default \code{"exposures"}) Set to get the missings: \code{"exposures"} or \code{"phenotypes"}}

\item{output}{\code{character} (default \code{"n"}) Get missing number (\code{"n"}) or percentage (\code{"p"})}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
Vector of named numerics, where the name corresponds to the variable and the associated numeric corresponds
to the missing (number or percentage, dependeing on the \code{output} parameter)
}
\description{
Obtain the number (or percentage) of missings of for the exposures or phenotypes of an ExposomeSet object
}
\examples{
\dontrun{Refer to the package Vignette for examples.}
}
