% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.transform.R
\name{ds.transform}
\alias{ds.transform}
\title{Apply a transformation to the exposure of an ExposomeSet}
\usage{
ds.transform(Set, exposure, method, new.obj = NULL, datasources = NULL)
}
\arguments{
\item{Set}{`character` Exposomeset on the server which exposures will be transformed}

\item{exposure}{`character` Name of the exposure to be transformed}

\item{method}{`character` Function to be applied to the exposure}

\item{new.obj}{`character` (default `NULL`) If provided, name of the new object that will
be created. If `NULL`, the input `Set` will be overwritten with the results}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}
}
\value{
This function does not have an output. It creates an Exposome Set object on the study server.
}
\description{
Apply a transformation to the exposure of an ExposomeSet
}
