% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_corPlot.R
\name{matrix_corPlotEXP}
\alias{matrix_corPlotEXP}
\title{Draws correlation matrix plot}
\usage{
matrix_corPlotEXP(corr, desc, cex.exposures = 0.5, cex.family = 0.55)
}
\arguments{
\item{corr}{Correlation}

\item{desc}{Fams}

\item{cex.exposures}{Alpha}

\item{cex.family}{Alpha}
}
\description{
This is a internal function, the code itself is taken from rexposome::.correlation_matrix_plot
}
