% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.exposome_HCPC.R
\name{ds.exposome_HCPC}
\alias{ds.exposome_HCPC}
\title{Title}
\usage{
ds.exposome_HCPC(object, nb.clust = -1, datasources = NULL)
}
\arguments{
\item{object}{}

\item{nb.clust}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
