% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.exposome_pca.R
\name{ds.exposome_pca}
\alias{ds.exposome_pca}
\title{Principal components analysis of an Exposome Set}
\usage{
ds.exposome_pca(
  Set,
  fam = NULL,
  scale = TRUE,
  type = "meta",
  pca = TRUE,
  npc = 10L,
  datasources = NULL
)
}
\arguments{
\item{Set}{\code{character} Name of the ExposomeSet on the study server}

\item{fam}{\code{character vector} (default \code{NULL}) Families to subset the ExposomeSet}

\item{scale}{\code{bool} (default \code{TRUE}) If \code{TRUE} the ExposomeSet wil be scaled, if \code{FALSE} 
it will not be scaled. It is always advisable to scale the data when performing a PCA.}

\item{type}{\code{character}  (default \code{"meta"}) Type of analysis, if meta-analysis (\code{"meta"}), a PCA to 
each study server will be performed. If pooled-analysis (\code{"pooled"}) a pooled methodology will be performed.}

\item{pca}{\code{bool} (default \code{TRUE}) If TRUE perform PCA (only numerical variables), 
if FALSE FAMD (numerical and categorical). This argument only affects \code{type = "meta"}, the pooled methodology 
does not have the option of performing a FAMD analysis.}

\item{npc}{\code{integer} (default \code{10L}) Number of principal components to compute.}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}
}
\value{
This function does not have an output. It creates a object on the study server named \code{ds.exposome_pca.Results},
this object can be passed to the \code{\link{ds.exposome_pca_plot}} to visualize the results of the PCA.
}
\description{
Performs a non-disclosive PCA given an Exposome Set on the study server, 
the Exposome Set can be subsetted by families to perform the PCA
}
\details{
The pooled PCA uses the block method for SVD decomposition. More information can be 
found on https://arxiv.org/pdf/1907.07410.pdf
}
\examples{
\dontrun{Refer to the package Vignette for examples.}
}
