% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.exposome_pca_plot.R
\name{ds.exposome_pca_plot}
\alias{ds.exposome_pca_plot}
\title{Title}
\usage{
ds.exposome_pca_plot(
  pca_object,
  set = "all",
  labels = FALSE,
  phenotype = NA,
  method,
  k,
  noise,
  datasources = NULL
)
}
\arguments{
\item{pca_object}{}

\item{set}{}

\item{phenotype}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
