% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.exposome_scale_exposures.R
\name{ds.exposome_scale_exposures}
\alias{ds.exposome_scale_exposures}
\title{Title}
\usage{
ds.exposome_scale_exposures(
  Set,
  method = c("combined", "split"),
  new.obj = NULL,
  datasources = NULL
)
}
\arguments{
\item{Set}{}

\item{new.obj}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
