% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.extractExposomeSetFromImputedSet.R
\name{ds.extractExposomeSetFromImputedSet}
\alias{ds.extractExposomeSetFromImputedSet}
\title{Title}
\usage{
ds.extractExposomeSetFromImputedSet(
  x,
  imputation = 1L,
  newobj.name = NULL,
  datasources = NULL
)
}
\arguments{
\item{x}{}

\item{imputation}{}

\item{datasources}{}
}
\value{

}
\description{
Title
}
