% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.get_exposure_from_geo.R
\name{ds.get_exposure_from_geo}
\alias{ds.get_exposure_from_geo}
\title{Title}
\usage{
ds.get_exposure_from_geo(
  lat,
  lon,
  exposure,
  exposure_name,
  clinical,
  clinical_lat_var,
  clinical_lon_var,
  clinical_id_var,
  exposures = NULL,
  exposures_id_var = NULL,
  new.obj = "exposures",
  datasources = NULL
)
}
\arguments{
\item{new.obj}{}

\item{datasources}{}

\item{nc}{}

\item{varid}{}
}
\description{
Title
}
