% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.ncatt_get.R
\name{ds.ncatt_get}
\alias{ds.ncatt_get}
\title{Title}
\usage{
ds.ncatt_get(nc, varid, attname, new.obj = "ncatt", datasources = NULL)
}
\arguments{
\item{nc}{}

\item{varid}{}

\item{new.obj}{}

\item{datasources}{}
}
\description{
Title
}
