% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.ncvar_get.R
\name{ds.ncvar_get}
\alias{ds.ncvar_get}
\title{Read data from netCDF file}
\usage{
ds.ncvar_get(nc, varid, new.obj = "ncvar", datasources = NULL)
}
\arguments{
\item{nc}{(Object server name of: )`character`
An object of class ncdf4 (loaded from a NetCDF resource), 
indicating what file to read from.}

\item{varid}{`character` What variable to read the data from.}
}
\value{
Creates object on the server
}
\description{
Reads data from an existing netCDF file.
}
