% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.NetCDF_fillvalue_matrix.R
\name{ds.NetCDF_fillvalue_matrix}
\alias{ds.NetCDF_fillvalue_matrix}
\title{Title}
\usage{
ds.NetCDF_fillvalue_matrix(mat, fillvalue, new.obj = "mat", datasources = NULL)
}
\arguments{
\item{new.obj}{}

\item{datasources}{}

\item{nc}{}

\item{varid}{}
}
\description{
Title
}
