% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.pca.R
\name{ds.pca}
\alias{ds.pca}
\title{Principal components analysis}
\usage{
ds.pca(tab, standar = TRUE, datasources = NULL)
}
\arguments{
\item{tab}{\code{character} Name of the data frame on the study server (can be of the form \code{table[, 1:10]})}

\item{standar}{\code{bool} Whether the values will be normalized prior the analysis (\code{TRUE}) or not (\code{FALSE})
Default \code{TRUE}}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}
}
\value{
A \code{data.frame} with principal components and variables
}
\description{
Performs a non-disclosive PCA given a data frame
}
\examples{
\dontrun{Refer to the package Vignette for examples.}
}
