% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.plotFamily.R
\name{ds.plotFamily}
\alias{ds.plotFamily}
\title{Draw boxplot of a family of an Exposome Set of the server side}
\usage{
ds.plotFamily(
  x,
  family,
  group = NULL,
  group2 = NULL,
  scatter = FALSE,
  na.omit = TRUE,
  method,
  k,
  noise,
  datasources = NULL
)
}
\arguments{
\item{x}{\code{character} Name of the Exposome Set on the server side}

\item{family}{\code{character} Name of the familty that will be drawn.  \coe{"all"} is to plot all the families
on a single mosaic (no grouping is used for the mosaic plot).}

\item{group}{\code{character} If set it displays the family grouped by the given phenotype}

\item{group2}{\code{character} If set it displays the family grouped by the given phenotype}

\item{scatter}{\code{bool} (default \code{TRUE}) If the family to be plotted is continuous, the samples will be shown}

\item{na.omit}{\code{bool} (default \code{TRUE}) Do not show NA values}

\item{method}{\code{numeric} either 1 or 2. If the user selects the deterministic method in the 
client side function the method.inticator is set to 1 while if the user selects the probabilistic method this argument is set to 2.}

\item{k}{\code{numeric} the number of the nearest neghbours for which their centroid is calculated}

\item{noise}{\code{numeric} the percentage of the initial variance that is used as the variance of}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
The function returns a ggplot object when a single family is selected and returns NULL for the mosaic (it renders
the plot on the current graphic device of the session)
}
\description{
Draw a boxplot for a single family or a mosaic of boxplots for all the families (only numeric families)
}
\examples{
\dontrun{Refer to the package Vignette for examples.}

}
