% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.plotMissings.R
\name{ds.plotMissings}
\alias{ds.plotMissings}
\title{Draw plot with percentage of missings}
\usage{
ds.plotMissings(exp, set = "exposures", datasources = NULL)
}
\arguments{
\item{exp}{\code{character} Name of the Exposome Set on the server side}

\item{set}{\code{character} (default \code{"exposures"}) Set to get the missings plot: \code{"exposures"} or \code{"phenotypes"}}

\item{datasources}{a list of \code{\link{DSConnection-class}} (default \code{NULL}) objects obtained after login}
}
\value{
Object of class \code{ggplot}, calling it will render the actual plot
}
\description{
Create a plot with the percentage of missings for the exposures and phenotypes of an 
ExposomeSet object
}
\examples{
\dontrun{Refer to the package Vignette for examples.}
}
